requirements.txt
aiogram==3.13.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==10.4.0
sqlalchemy==2.0.36
alembic==1.13.2
python-dateutil==2.9.0.post0
